requirements.txt
openai>=1.0.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
schedule>=1.2.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
aiohttp>=3.8.0
asyncio>=3.4.3
scikit-learn>=1.3.0
matplotlib>=3.7.0
plotly>=5.15.0
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
httpx>=0.24.0
tweepy>=4.14.0
ccxt>=4.0.0
python-telegram-bot>=20.0
redis>=4.6.0
celery>=5.3.0
psycopg2-binary>=2.9.0